feat(ServiceWorker):fix 2
