+++ src/app/app.component.html
@@ -1,1 +1,39 @@
 <router-outlet></router-outlet>
+
+<div bsModal #notificationModal="bs-modal" class="modal fade" tabindex="-1" role="dialog"
+     aria-labelledby="Label"
+     aria-hidden="true">
+  <div class="modal-dialog" role="document">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h4 class="modal-title">{{ notificationModalTitle }}</h4>
+        <button type="button" class="close" style="cursor:pointer"
+                (click)="notificationModal.hide()"
+                aria-label="Close">
+          <span aria-hidden="true">&times;</span>
+        </button>
+      </div>
+      <div class="modal-body">
+        <div class="alert alert-{{ notificationModalType }}" role="alert">
+          {{ notificationModalMessage }}
+
+        </div>
+      </div>
+      <div class="modal-footer">
+        <button
+                type="button"
+                (click)="notificationModal.hide()"
+                class="btn  btn-secondary"> Close
+        </button>
+          <button type="button"
+                  (click)="notificationModal.hide();reloadSite();" type="reset"
+                  class="btn  btn-secondary"> Reload
+          </button>
+      </div>
+    </div><!-- /.modal-content -->
+  </div><!-- /.modal-dialog -->
+</div><!-- /.modal -->
+
+
+<button  (click)="openNotificationModal()">Show snack-bar</button>
+
