+++ src/app/app.component.ts
@@ -1,8 +1,11 @@
 import {AfterViewInit, Component, OnInit, ViewChild} from '@angular/core';
 import {Http, RequestOptions, XHRBackend} from '@angular/http';
 import {ModalDirective} from 'ngx-bootstrap';
-import {UpdateService} from './update.service';
+import {SwUpdate} from '@angular/service-worker';
 import {Angulartics2Piwik} from 'angulartics2/piwik';
+import {ApplicationRef} from '@angular/core';
+import {concat, interval} from "rxjs";
+import {first} from 'rxjs/operators';
 
 /**
  * App component.
@@ -10,19 +13,43 @@ import {Angulartics2Piwik} from 'angulartics2/piwik';
 @Component({
     selector: 'body',
     templateUrl: 'app.component.html',
-    providers: [UpdateService, {
+    providers: [{
         provide: Http,
         deps: [XHRBackend, RequestOptions, AppComponent]
     }]
 })
 export class AppComponent implements AfterViewInit, OnInit {
 
-    @ViewChild('timeoutModal') modal: ModalDirective;
+    notificationModalTitle = 'Update available';
+    notificationModalMessage = 'A new update is available. Please reload the site to use the new version of the portal.';
+    notificationModalType = 'info';
+
+    @ViewChild('notificationModal') modal: ModalDirective;
 
 
-    constructor(private update: UpdateService, private angulartics2Piwik: Angulartics2Piwik) {
+    constructor(private appRef: ApplicationRef, private swUpdate: SwUpdate, private angulartics2Piwik: Angulartics2Piwik) {
+        const isStable = appRef.isStable.pipe(first(isStable => isStable === true));
+        const intervalTime = interval(60 * 1000);
+        const checkUpdatesInIntervall = concat(isStable, intervalTime);
+        checkUpdatesInIntervall.subscribe(() => this.swUpdate.checkForUpdate().then(() => {
+            this.swUpdate.available.subscribe(evt => {
+                this.openNotificationModal()
+
+            })
+        }))
+
 
     }
+
+    reloadSite() {
+        window.location.reload()
+    }
+
+    openNotificationModal() {
+
+        this.modal.show()
+    }
+
 
     ngOnInit(): void {
         this.angulartics2Piwik.startTracking();
